refactor: Restructure project modules to allow for compiling in different java versions
