+++ SkyWarsReloadedCore/pom.xml
@@ -11,15 +11,13 @@
 
     <artifactId>SkyWarsReloadedCore</artifactId>
     <packaging>jar</packaging>
-    <version>${project.parent.version}</version>
+    <version>${revision}</version>
 
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <outputDirectory>./target/</outputDirectory>
     </properties>
-
 
     <repositories>
         <repository>
@@ -114,12 +112,7 @@
             <scope>provided</scope>
         </dependency>
         <!-- SLIME WORLDS: ASWM DEPENDENCY -->
-        <dependency>
-            <groupId>com.grinderwolf</groupId>
-            <artifactId>AdvancedSlimeWorldManager</artifactId>
-            <version>2.8.0</version>
-            <scope>provided</scope>
-        </dependency>
+
         <!-- SLIME WORLDS: USE THESE DEPENDENCIES TO GET LEGACY SWM SUPPORT IN THE IDEA DURING DEVELOPMENT-->
         <!--<dependency>
             <groupId>com.grinderwolf</groupId>
