+++ SkyWarsReloadedCore/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: Skywars
 main: com.walrusone.skywarsreloaded.SkyWarsReloaded
-version: ${project.version}
+version: ${revision}
 api-version: 1.13
 website: https://gcnt.net/
 author: GaagjesCraft Network Team (GCNT)
