+++ pom.xml
@@ -10,7 +10,6 @@
     <version>${revision}</version>
 
     <properties>
-
         <revision>5.6.32</revision>
 
         <bukkit.version>1.12.2</bukkit.version>
@@ -22,6 +21,7 @@
     <modules>
         <module>SkyWarsReloadedCore</module>
         <module>slime-asp</module>
+        <module>slime-legacy</module>
         <module>v1_8_R3</module>
         <module>v1_9_R1</module>
         <module>v1_9_R2</module>
@@ -37,18 +37,125 @@
         <module>v1_17_R1</module>
         <module>v1_18_R2</module>
         <module>v1_19_R1</module>
-        <module>SW-Packager</module>
+        <module>skywars-assembly</module>
     </modules>
+
+    <profiles>
+        <profile>
+            <id>java8</id>
+            <properties>
+                <maven.compiler.source>1.8</maven.compiler.source>
+                <maven.compiler.target>1.8</maven.compiler.target>
+            </properties>
+            <activation>
+                <jdk>1.8</jdk>
+            </activation>
+            <modules>
+                <module>SkyWarsReloadedCore</module>
+                <module>v1_8_R3</module>
+                <module>v1_9_R1</module>
+                <module>v1_9_R2</module>
+                <module>v1_10_R1</module>
+                <module>v1_11_R1</module>
+                <module>v1_12_R1</module>
+                <module>v1_13_R2</module>
+                <module>v1_14_R1</module>
+                <module>v1_15_R1</module>
+                <module>v1_16_R1</module>
+                <module>v1_16_R2</module>
+                <module>v1_16_R3</module>
+            </modules>
+        </profile>
+        <profile>
+            <id>java16</id>
+            <properties>
+                <maven.compiler.source>16</maven.compiler.source>
+                <maven.compiler.target>16</maven.compiler.target>
+            </properties>
+            <activation>
+                <jdk>16</jdk>
+            </activation>
+            <modules>
+                <module>v1_17_R1</module>
+            </modules>
+        </profile>
+        <profile>
+            <id>java17</id>
+            <properties>
+                <maven.compiler.source>17</maven.compiler.source>
+                <maven.compiler.target>17</maven.compiler.target>
+            </properties>
+            <activation>
+                <jdk>17</jdk>
+            </activation>
+            <modules>
+                <module>slime-asp</module>
+                <module>slime-legacy</module>
+                <module>v1_18_R2</module>
+                <module>v1_19_R1</module>
+            </modules>
+        </profile>
+    </profiles>
+
+    <repositories>
+        <repository>
+            <id>bukkitmc</id>
+            <url>https://nexuslite.gcnt.net/repos/bukkit/</url>
+        </repository>
+    </repositories>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>com.walrusone</groupId>
+                <artifactId>SkyWarsReloadedCore</artifactId>
+                <version>${revision}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
     <build>
         <plugins>
             <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-toolchains-plugin</artifactId>
+                <version>3.1.0</version>
+                <executions>
+                    <execution>
+                        <phase>validate</phase>
+                        <goals>
+                            <goal>toolchain</goal>
+                        </goals>
+                    </execution>
+                </executions>
+                <configuration>
+                    <toolchains>
+                        <jdk>
+                            <version>${maven.compiler.source}</version>
+                        </jdk>
+                    </toolchains>
+                </configuration>
+            </plugin>
+            <!-- Maven Compiler Plugin -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.0</version>
+                <configuration>
+                    <source>${maven.compiler.source}</source>
+                    <target>${maven.compiler.target}</target>
+                    <jdkToolchain>
+                        <type>jdk</type>
+                        <version>${maven.compiler.target}</version>
+                    </jdkToolchain>
+                </configuration>
+            </plugin>
+
+            <!-- Optional Flatten plugin, if you're still using it -->
+            <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>flatten-maven-plugin</artifactId>
                 <version>1.0.0</version>
-                <configuration>
-                    <updatePomFile>true</updatePomFile>
-                </configuration>
                 <executions>
                     <execution>
                         <id>flatten</id>
@@ -66,31 +173,6 @@
                     </execution>
                 </executions>
             </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.0</version>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-deploy-plugin</artifactId>
-                <version>2.8.1</version>
-                <configuration>
-                    <skip>true</skip>
-                </configuration>
-            </plugin>
         </plugins>
     </build>
-
-    <distributionManagement>
-        <repository>
-            <id>swr-nexus</id>
-            <name>SkywarsReloaded Nexus</name>
-            <url>https://nexus.gcnt.net/repository/gaagjescraft/</url>
-        </repository>
-    </distributionManagement>
 </project>
