+++ v1_18_R2/src/main/java/com/walrusone/skywarsreloaded/nms/v1_18_R2/NMSHandler.java
@@ -1,17 +1,13 @@
 package com.walrusone.skywarsreloaded.nms.v1_18_R2;
 
-import net.minecraft.core.BlockPosition;
-import net.minecraft.server.level.WorldServer;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.block.Biome;
 import org.bukkit.block.Block;
-import org.bukkit.block.EnderChest;
 import org.bukkit.entity.*;
 import org.bukkit.generator.ChunkGenerator;
-import org.bukkit.scoreboard.Scoreboard;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Random;
